requirements.txt
# --- LLM and TTS Dependencies ---
openai
elevenlabs
google-genai
gtts

# --- Video Processing (moviepy and its core dependencies) ---
moviepy==1.0.3
numpy==1.26.4
imageio==2.37.0
imageio-ffmpeg==0.6.0
pillow==11.0.0
proglog==0.1.12

# --- Other Utilities ---
python-dotenv==1.0.1
requests==2.32.5
certifi==2025.10.5
charset-normalizer==3.4.3
decorator
idna==3.10
tqdm==4.66.5
urllib3==2.5.0
pydub==0.25.1
